https://github.com/pendulum-chain/spacewalk/pull/335#discussion_r1168362794, https://github.com/pendulum-chain/spacewalk/pull/335#discussion_r1168364850
